Refactored assignments.py and put code in other files
